04-Prompt-Engineering/projects/5_evaluation_prompts/5-langfuse/requirements.txt
langfuse>=2.0.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
